add conformance test for deit-small
